+++ R/model.R
@@ -37,7 +37,7 @@
 #' ex  <- function(f) system.file("extdata", f, package = "msamodel")
 #' wt  <- set_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca",
 #'                       model = "ming_wall", d_max = 10.5, frustrated = FALSE)
-#' spm <- generate_spm(wt, seed = 1024)
+#' spm <- generate_spm(wt, ensemble = 1L)
 #' pfix_msa(spm$energy_data$ddg, spm$energy_data$ddgact, a1 = 1, a2 = 1)
 #' }
 #' @export
@@ -265,7 +265,7 @@ unimplemented_metric_message <- function(metric) {
 #' wt  <- set_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca",
 #'                       model = "ming_wall", d_max = 10.5, frustrated = FALSE)
 #' act <- readr::read_csv(ex("znb_active_site.csv"))
-#' spm <- generate_spm(wt, pdb_site_active = act$pdb_site, seed = 1024)
+#' spm <- generate_spm(wt, pdb_site_active = act$pdb_site, ensemble = 1L)
 #' calculate_profiles(spm, a1 = 1, a2 = 1, metric = "nlrmsd")$site
 #' }
 #' @export
@@ -327,7 +327,7 @@ calculate_profiles <- function(spm, a1, a2, metric = c("lrmsd", "nlrmsd")) {
 #' wt  <- set_enm(bio3d::read.pdb(ex("1znb_A.pdb")), node = "ca",
 #'                       model = "ming_wall", d_max = 10.5, frustrated = FALSE)
 #' act <- readr::read_csv(ex("znb_active_site.csv"))
-#' spm <- generate_spm(wt, pdb_site_active = act$pdb_site, seed = 1024)
+#' spm <- generate_spm(wt, pdb_site_active = act$pdb_site, ensemble = 1L)
 #' calculate_decomposition(spm, a1 = 1, a2 = 1, metric = "nlrmsd")$site
 #' }
 #' @export